Added Builder tests
